api_test_framework/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "api-test-framework"
version = "1.0.0"
description = "接口自动化测试框架"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "pytest>=7.4.3",
    "requests>=2.31.0",
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.0",
    "jsonschema>=4.20.0",
    "pydantic>=2.5.2",
    "faker>=22.0.0",
    "jmespath>=1.0.1",
    "cryptography>=41.0.7",
]

[project.optional-dependencies]
dev = [
    "pytest-cov>=4.1.0",
    "pytest-html>=4.1.1",
    "pytest-xdist>=3.5.0",
    "pytest-rerunfailures>=13.0",
    "allure-pytest>=2.13.2",
    "requests-mock>=1.11.0",
    "bandit>=1.7.6",
    "safety>=3.0.1",
    "black>=23.12.1",
    "isort>=5.13.2",
    "mypy>=1.8.0",
    "pylint>=3.0.3",
    "radon>=6.0.1",
]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --strict-config --tb=short"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "smoke: 冒烟测试",
    "regression: 回归测试",
    "security: 安全测试",
    "performance: 性能测试",
    "api: 接口测试",
    "unit: 单元测试",
    "slow: 慢速测试",
    "integration: 集成测试",
]

[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip_gitignore = true
skip = [".git", "__pycache__", "build", "dist"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
check_untyped_defs = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[[tool.mypy.overrides]]
module = "fixtures.*"
disallow_untyped_defs = false
